requirements.txt
biopython==1.66
cycler==0.10.0
decorator==4.0.9
ete2==2.3.10
matplotlib==1.5.1
networkx==1.11
numpy==1.11.0
pyparsing==2.1.1
python-dateutil==2.5.3
pytz==2016.4
six==1.10.0
wsgiref==0.1.2
